+++ tests/cases/test_gui_joystick_constants.py
@@ -1,0 +1,17 @@
+# expect:
+# 512
+# 0
+# 1
+# 2
+# 1536
+# 1539
+# 1540
+import _gui_sdl
+
+print(_gui_sdl.INIT_JOYSTICK)
+print(_gui_sdl.FLIP_NONE)
+print(_gui_sdl.FLIP_HORIZONTAL)
+print(_gui_sdl.FLIP_VERTICAL)
+print(_gui_sdl.EVENT_JOYAXISMOTION)
+print(_gui_sdl.EVENT_JOYBUTTONDOWN)
+print(_gui_sdl.EVENT_JOYBUTTONUP)